# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Employee MCP Server",
    "description": "Bridge API for Employee Management System. Provides AI-agent compatible tools to Create, Read, Update, and Delete employee records.",
    "version": "1.0"
  },
  "paths": {
    "/employees": {
      "get": {
        "tags": [
          "Employee READ Operation"
        ],
        "summary": "Retrieve all employee records",
        "description": "Tool to fetch the full list of all employees in the system. Use this when the user asks 'who works here', 'list all employees', or needs to find a specific person's ID before updating them.",
        "operationId": "GetAllEmployees",
        "responses": {
          "200": {
            "description": "Employee list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EmployeeResponse"
                  },
                  "type": "array",
                  "title": "Response Getallemployees"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Employee CREATE Operation"
        ],
        "summary": "Create a new employee record",
        "description": "Tool to create a new employee in the database. Use this when a user asks to onboard, add, or create a new employee. Requires name, role, and department.",
        "operationId": "CreateEmployee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Employee created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/employees/{emp_id}": {
      "put": {
        "tags": [
          "Employee UPDATE operation"
        ],
        "summary": "Update an employee record",
        "description": "Tool to modify an existing employee's details (like name, role, or department). You must provide the numeric emp_id in the path alongside the new payload. Use GetAllEmployees first if you don't know the emp_id.",
        "operationId": "UpdateEmployee",
        "parameters": [
          {
            "name": "emp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Emp Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Employee DELETE Operation"
        ],
        "summary": "Delete an employee record",
        "description": "Tool to remove an employee from the system. Requires the numeric emp_id. Use this when a user asks to terminate, remove, or delete an employee.",
        "operationId": "DeleteEmployee",
        "parameters": [
          {
            "name": "emp_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Emp Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EmployeeInput": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The full name of the employee, e.g., 'Jane Doe'."
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "The job title or role, e.g., 'Software Engineer' or 'Manager'."
          },
          "department": {
            "type": "string",
            "title": "Department",
            "description": "The department name, e.g., 'Engineering' or 'HR'."
          }
        },
        "type": "object",
        "required": [
          "name",
          "role",
          "department"
        ],
        "title": "EmployeeInput"
      },
      "EmployeeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "department": {
            "type": "string",
            "title": "Department"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "department"
        ],
        "title": "EmployeeResponse"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      }
    }
  },
  "servers": [
    {
      "url": "https://mcp-server.whiteforest-3a1a66f2.eastus.azurecontainerapps.io",
      "description": "MCP Server"
    }
  ]
}